cookies/cookie-get-srcdoc.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
    var t = async_test("document.cookie access in a srcdoc");
    window.onmessage = function(ev) {
      // Surprisingly, the legacy behavior here is to return true; this actually
      // does match the spec definition since false is supposed to be returned
      // when a document.cookie write is ignored --- and here it would throw
      // a security exception, not be ignored.
      assert_equals(ev.data, "???");
      t.done();
    }

    t.add_cleanup(function() {
      document.cookie = "srcdoc_test=1;max-age=0;";
    });

    t.step(function() {
      var iframe = document.createElement("iframe");
      document.cookie = "srcdoc_test=1";
      iframe.srcdoc = "<scr" + "ipt>" +
            "window.onmessage = function() {" +
            "  parent.postMessage(document.cookie, '*'); " +
            "}</scr" + "ipt>";
      iframe.onload = function() {
        iframe.contentWindow.postMessage({}, "*");
      }
      document.body.appendChild(iframe);
    });
  </script>
</body>
